Consultas parciales necesarias para empresa, partidas, registros de partidas y otros, agregadas

+++ OptiFinance_System/database/models/Registro.cs
@@ -3,8 +3,8 @@
 public class Registro
 {
     public long Id { get; set; }
-    public double Debe { get; set; }
-    public double Haber { get; set; }
+    public decimal Debe { get; set; }
+    public decimal Haber { get; set; }
     public Cuenta Cuenta { get; set; } = new Cuenta();
     public Partida Partida { get; set; } = new Partida();
